feat: consolidate multiple utils files
