+++ src/fides/api/api/v1/endpoints/admin.py
@@ -4,11 +4,11 @@ from typing import Dict
 from fastapi import Security
 
 from fides.api.api.v1 import scope_registry
-from fides.api.api.v1.endpoints.util import API_PREFIX
+from fides.api.api.v1.endpoints.utils import API_PREFIX
 from fides.api.db import database
 from fides.api.oauth.utils import verify_oauth_client_prod
-from fides.api.utils import errors
-from fides.api.utils.api_router import APIRouter
+from fides.api.util import errors
+from fides.api.util.api_router import APIRouter
 from fides.core.config import CONFIG
 
 ADMIN_ROUTER = APIRouter(prefix=API_PREFIX, tags=["Admin"])
