+++ src/fides/api/api/v1/endpoints/generate.py
@@ -11,12 +11,12 @@ from pydantic import BaseModel, root_validator
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from fides.api.api.v1 import scope_registry
-from fides.api.api.v1.endpoints.util import API_PREFIX
+from fides.api.api.v1.endpoints.utils import API_PREFIX
 from fides.api.db.crud import get_resource
 from fides.api.db.ctl_session import get_async_db
 from fides.api.models.sql_models import sql_model_map  # type: ignore[attr-defined]
 from fides.api.oauth.utils import verify_oauth_client_prod
-from fides.api.utils.api_router import APIRouter
+from fides.api.util.api_router import APIRouter
 from fides.connectors.models import (
     AWSConfig,
     BigQueryConfig,
