+++ src/fides/api/api/v1/endpoints/health.py
@@ -11,9 +11,9 @@ from fides.api.api.deps import get_db
 from fides.api.common_exceptions import RedisConnectionError
 from fides.api.db.database import DatabaseHealth, get_db_health
 from fides.api.tasks import celery_app, get_worker_ids
+from fides.api.util.api_router import APIRouter
 from fides.api.util.cache import get_cache
 from fides.api.util.logger import Pii
-from fides.api.utils.api_router import APIRouter
 from fides.core.config import CONFIG
 
 CacheHealth = Literal["healthy", "unhealthy", "no cache configured"]
