+++ src/fides/api/api/v1/endpoints/router_factory.py
@@ -14,7 +14,7 @@ from pydantic import ValidationError as PydanticValidationError
 from sqlalchemy.ext.asyncio import AsyncSession
 from starlette.status import HTTP_422_UNPROCESSABLE_ENTITY
 
-from fides.api.api.v1.endpoints.util import (
+from fides.api.api.v1.endpoints.utils import (
     API_PREFIX,
     CLI_SCOPE_PREFIX_MAPPING,
     forbid_if_default,
@@ -38,8 +38,8 @@ from fides.api.models.sql_models import (
 )
 from fides.api.oauth.utils import verify_oauth_client_prod
 from fides.api.schemas.dataset import validate_data_categories_against_db
-from fides.api.utils import errors
-from fides.api.utils.api_router import APIRouter
+from fides.api.util import errors
+from fides.api.util.api_router import APIRouter
 
 
 async def get_data_categories_from_db(async_session: AsyncSession) -> List[FidesKey]:
