+++ src/fides/api/api/v1/endpoints/validate.py
@@ -8,9 +8,9 @@ from fastapi import Response, Security, status
 from pydantic import BaseModel
 
 from fides.api.api.v1 import scope_registry
-from fides.api.api.v1.endpoints.util import API_PREFIX
+from fides.api.api.v1.endpoints.utils import API_PREFIX
 from fides.api.oauth.utils import verify_oauth_client_prod
-from fides.api.utils.api_router import APIRouter
+from fides.api.util.api_router import APIRouter
 from fides.connectors.models import (
     AWSConfig,
     BigQueryConfig,
