+++ src/fides/api/api/v1/endpoints/view.py
@@ -10,7 +10,7 @@ from fides.api.db.crud import list_resource
 from fides.api.db.ctl_session import get_async_db
 from fides.api.models.sql_models import Evaluation  # type: ignore[attr-defined]
 from fides.api.oauth.utils import verify_oauth_client_prod
-from fides.api.utils.api_router import APIRouter
+from fides.api.util.api_router import APIRouter
 
 router = APIRouter(
     tags=["View"],
