+++ src/fides/api/app_setup.py
@@ -35,14 +35,14 @@ from fides.api.util.consent_util import (
     load_default_experience_configs_on_startup,
     load_default_notices_on_startup,
 )
+from fides.api.util.errors import FidesError
+from fides.api.util.logger import setup as setup_logging
 from fides.api.util.system_manager_oauth_util import (
     get_system_fides_key,
     get_system_schema,
     verify_oauth_client_for_system_from_fides_key_cli,
     verify_oauth_client_for_system_from_request_body_cli,
 )
-from fides.api.utils.errors import FidesError
-from fides.api.utils.logger import setup as setup_logging
 from fides.core.config import CONFIG
 
 VERSION = fides.__version__
