+++ src/fides/api/db/database.py
@@ -15,7 +15,7 @@ from sqlalchemy_utils.types.encrypted.encrypted_type import InvalidCiphertextErr
 from fides.api.db.base import Base  # type: ignore[attr-defined]
 from fides.api.db.ctl_session import async_session
 from fides.api.db.seed import load_default_resources, load_samples
-from fides.api.utils.errors import get_full_exception_name
+from fides.api.util.errors import get_full_exception_name
 from fides.core.utils import get_db_engine
 
 DatabaseHealth = Literal["healthy", "unhealthy", "needs migration"]
