+++ src/fides/api/service/connectors/fides/fides_client.py
@@ -16,8 +16,8 @@ from fides.api.schemas.redis_cache import Identity
 from fides.api.schemas.user import UserLogin
 from fides.api.service.privacy_request.request_service import poll_server_for_completion
 from fides.api.util.collection_util import Row
+from fides.api.util.errors import FidesError
 from fides.api.util.wrappers import sync
-from fides.api.utils.errors import FidesError
 
 COMPLETION_STATUSES = [
     PrivacyRequestStatus.complete,
