+++ tests/ctl/api/test_generate.py
@@ -7,7 +7,7 @@ import pytest
 from starlette.testclient import TestClient
 
 from fides.api.api.v1.endpoints.generate import GenerateResponse
-from fides.api.api.v1.endpoints.util import API_PREFIX
+from fides.api.api.v1.endpoints.utils import API_PREFIX
 from fides.core.config import FidesConfig
 
 EXTERNAL_CONFIG_BODY = {
