+++ tests/ctl/api/test_validate.py
@@ -6,7 +6,7 @@ from os import getenv
 import pytest
 from starlette.testclient import TestClient
 
-from fides.api.api.v1.endpoints.util import API_PREFIX
+from fides.api.api.v1.endpoints.utils import API_PREFIX
 from fides.api.api.v1.endpoints.validate import ValidateResponse
 from fides.core.config import FidesConfig
