+++ tests/ops/service/connectors/fides/test_fides_client.py
@@ -6,7 +6,7 @@ from httpx import AsyncClient, Client, HTTPStatusError
 
 from fides.api.models.privacy_request import PrivacyRequest, PrivacyRequestStatus
 from fides.api.service.connectors.fides.fides_client import FidesClient
-from fides.api.utils.errors import FidesError
+from fides.api.util.errors import FidesError
 
 SAMPLE_TOKEN = "SOME_TOKEN"
